Refactor governance orchestration into service clients
