+++ docs/implementations/integration/spark-dlt.md
@@ -11,7 +11,7 @@ dc43 keeps governance logic decoupled from runtime execution. The integration la
 
 ```mermaid
 flowchart TD
-    ContractStore["Contract store"] --> IO["dc43.components.integration.spark_io"]
+    ContractStore["Contract store"] --> IO["dc43.integration.spark_io"]
     IO --> Spark["Spark Jobs / DLT"]
     IO --> ContractMgr["Data contract manager"]
     IO --> Governance["Governance service"]
@@ -23,17 +23,17 @@ flowchart TD
 
 ## Spark & Delta Helpers
 
-The canonical implementation lives in [`src/dc43/components/integration`](../../src/dc43/components/integration):
+The canonical implementation lives in [`src/dc43/integration`](../../src/dc43/integration):
 
 * `spark_io.py` — High-level `read_with_contract` and `write_with_contract` wrappers for Spark DataFrames along with dataset resolution helpers.
 * `dlt_helpers.py` — Functions to translate ODCS expectations into Delta Live Tables expectations.
 * [`dc43.components.data_quality.integration`](../../src/dc43/components/data_quality/integration/__init__.py) — Schema snapshots, expectation predicates, and metric builders used by adapters.
-* [`dc43.components.governance_service`](../../src/dc43/components/governance_service) — Coordination service that links contracts, evaluates observations, and persists drafts.
+* [`dc43.services.governance`](../../src/dc43/services/governance) — Coordination service that links contracts, evaluates observations, and persists drafts.
 
 Pipelines typically import these helpers directly:
 
 ```python
-from dc43.components.integration.spark_io import (
+from dc43.integration.spark_io import (
     read_with_contract,
     write_with_contract,
     ContractVersionLocator,
