+++ src/dc43/demo_app/pipeline.py
@@ -23,8 +23,8 @@ from dc43.demo_app.server import (
 )
 from dc43.components.data_quality.governance import DQStatus
 from dc43.components.data_quality.integration import attach_failed_expectations
-from dc43.components.governance_service import build_local_governance_service
-from dc43.components.integration.spark_io import (
+from dc43.services.governance.local import build_local_governance_service
+from dc43.integration.spark_io import (
     ContractFirstDatasetLocator,
     ContractVersionLocator,
     ReadStatusContext,
@@ -33,7 +33,7 @@ from dc43.components.integration.spark_io import (
     read_with_contract,
     write_with_contract,
 )
-from dc43.components.integration.violation_strategy import (
+from dc43.integration.violation_strategy import (
     NoOpWriteViolationStrategy,
     SplitWriteViolationStrategy,
     StrictWriteViolationStrategy,
