+++ src/dc43/integration/violation_strategy.py
