+++ src/dc43/services/__init__.py
@@ -1,0 +1,32 @@
+"""High-level service clients used by integration layers."""
+
+from .data_quality import DataQualityServiceClient, LocalDataQualityServiceClient, ObservationPayload
+from .governance import (
+    GovernanceCredentials,
+    GovernanceServiceClient,
+    LocalGovernanceService,
+    PipelineContext,
+    PipelineContextSpec,
+    QualityAssessment,
+    QualityDraftContext,
+    build_local_governance_service,
+    normalise_pipeline_context,
+)
+from .contracts import ContractServiceClient, LocalContractServiceClient
+
+__all__ = [
+    "DataQualityServiceClient",
+    "LocalDataQualityServiceClient",
+    "ObservationPayload",
+    "GovernanceCredentials",
+    "GovernanceServiceClient",
+    "LocalGovernanceService",
+    "PipelineContext",
+    "PipelineContextSpec",
+    "QualityAssessment",
+    "QualityDraftContext",
+    "build_local_governance_service",
+    "normalise_pipeline_context",
+    "ContractServiceClient",
+    "LocalContractServiceClient",
+]
