+++ src/dc43/services/contracts/__init__.py
@@ -1,0 +1,6 @@
+"""Client helpers for interacting with contract management services."""
+
+from .client import ContractServiceClient
+from .local import LocalContractServiceClient
+
+__all__ = ["ContractServiceClient", "LocalContractServiceClient"]
