+++ src/dc43/services/data_quality/__init__.py
@@ -1,0 +1,10 @@
+"""Client utilities for interacting with data-quality services."""
+
+from .client import DataQualityServiceClient, LocalDataQualityServiceClient
+from .models import ObservationPayload
+
+__all__ = [
+    "DataQualityServiceClient",
+    "LocalDataQualityServiceClient",
+    "ObservationPayload",
+]
