+++ src/dc43/services/data_quality/models.py
@@ -1,0 +1,18 @@
+"""Shared models for interacting with data-quality services."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Mapping, Optional
+
+
+@dataclass(slots=True)
+class ObservationPayload:
+    """Container describing cached observations for a dataset evaluation."""
+
+    metrics: Mapping[str, object]
+    schema: Optional[Mapping[str, Mapping[str, object]]] = None
+    reused: bool = False
+
+
+__all__ = ["ObservationPayload"]
