+++ src/dc43/services/governance/__init__.py
@@ -1,0 +1,24 @@
+"""Governance orchestration service abstractions and local stubs."""
+
+from .client import GovernanceServiceClient
+from .local import LocalGovernanceService, build_local_governance_service
+from .models import (
+    GovernanceCredentials,
+    PipelineContext,
+    PipelineContextSpec,
+    QualityAssessment,
+    QualityDraftContext,
+    normalise_pipeline_context,
+)
+
+__all__ = [
+    "GovernanceServiceClient",
+    "GovernanceCredentials",
+    "PipelineContext",
+    "PipelineContextSpec",
+    "QualityAssessment",
+    "QualityDraftContext",
+    "normalise_pipeline_context",
+    "LocalGovernanceService",
+    "build_local_governance_service",
+]
