+++ src/dc43/services/governance/client.py
@@ -1,0 +1,113 @@
+"""Client protocol for governance orchestration services."""
+
+from __future__ import annotations
+
+from typing import Callable, Mapping, Optional, Protocol, Sequence
+
+from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
+
+from dc43.components.data_quality.engine import ValidationResult
+from dc43.components.data_quality.governance import DQStatus
+from dc43.services.data_quality.models import ObservationPayload
+
+from .models import (
+    GovernanceCredentials,
+    PipelineContextSpec,
+    QualityAssessment,
+    QualityDraftContext,
+)
+
+
+class GovernanceServiceClient(Protocol):
+    """High-level coordination API exposed to integration layers."""
+
+    def configure_auth(
+        self,
+        credentials: GovernanceCredentials | Mapping[str, object] | str | None,
+    ) -> None:
+        ...
+
+    def evaluate_dataset(
+        self,
+        *,
+        contract_id: str,
+        contract_version: str,
+        dataset_id: str,
+        dataset_version: str,
+        validation: ValidationResult | None,
+        observations: Callable[[], ObservationPayload],
+        bump: str = "minor",
+        context: QualityDraftContext | None = None,
+        pipeline_context: PipelineContextSpec | None = None,
+        operation: str = "read",
+        draft_on_violation: bool = False,
+    ) -> QualityAssessment:
+        ...
+
+    def review_validation_outcome(
+        self,
+        *,
+        validation: ValidationResult,
+        base_contract: OpenDataContractStandard,
+        bump: str = "minor",
+        dataset_id: str | None = None,
+        dataset_version: str | None = None,
+        data_format: str | None = None,
+        dq_status: DQStatus | None = None,
+        dq_feedback: Mapping[str, object] | None = None,
+        context: QualityDraftContext | None = None,
+        pipeline_context: PipelineContextSpec | None = None,
+        draft_requested: bool = False,
+        operation: str | None = None,
+    ) -> Optional[OpenDataContractStandard]:
+        ...
+
+    def propose_draft(
+        self,
+        *,
+        validation: ValidationResult,
+        base_contract: OpenDataContractStandard,
+        bump: str = "minor",
+        context: QualityDraftContext | None = None,
+        pipeline_context: PipelineContextSpec | None = None,
+    ) -> OpenDataContractStandard:
+        ...
+
+    def get_status(
+        self,
+        *,
+        contract_id: str,
+        contract_version: str,
+        dataset_id: str,
+        dataset_version: str,
+    ) -> Optional[DQStatus]:
+        ...
+
+    def link_dataset_contract(
+        self,
+        *,
+        dataset_id: str,
+        dataset_version: str,
+        contract_id: str,
+        contract_version: str,
+    ) -> None:
+        ...
+
+    def get_linked_contract_version(
+        self,
+        *,
+        dataset_id: str,
+        dataset_version: Optional[str] = None,
+    ) -> Optional[str]:
+        ...
+
+    def get_pipeline_activity(
+        self,
+        *,
+        dataset_id: str,
+        dataset_version: Optional[str] = None,
+    ) -> Sequence[Mapping[str, object]]:
+        ...
+
+
+__all__ = ["GovernanceServiceClient"]
